Address code review feedback on the request-budget test suite
- Extract respond_ok_to_every_sub_request() and the MERCHANT_ID
  fixture into a shared RequestBudgetTestCase base class, removing
  the byte-for-byte duplicate between ProductUpsertRequestBudgetTest
  and DataSourceResolutionRequestBudgetTest.
- Remove the dead PROCESS_ITEM_HOOK constant from
  ProductStatusRefreshRequestBudgetTest (never referenced - the hook
  is propagated via $job->init()/schedule() at runtime).
- Rename that class's budget_sizes() data provider to
  page_configurations() to stop colliding by name with the
  differently-shaped provider in ProductUpsertRequestBudgetTest.
- Add the missing assert_client_never_calls_a_per_item_write_method()
  call to the two DataSourceResolutionRequestBudgetTest tests that
  only asserted get()/post() counts, for consistency with the
  data-provider tests in the same file.
